+++ src/main/java/com/mclegoman/perspective/mixin/client/zoom/MouseMixin.java
@@ -7,18 +7,14 @@
 
 package com.mclegoman.perspective.mixin.client.zoom;
 
-import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.mclegoman.perspective.client.config.ConfigHelper;
 import com.mclegoman.perspective.client.data.ClientData;
 import com.mclegoman.perspective.client.zoom.Zoom;
 import net.minecraft.client.Mouse;
-import net.minecraft.client.network.ClientPlayerEntity;
-import net.minecraft.util.math.MathHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(priority = 10000, value = Mouse.class)
@@ -56,11 +52,9 @@ public abstract class MouseMixin {
 	}
 	@Inject(method = "updateMouse", at = @At(value = "HEAD"))
 	private void perspective$updateMouse(double timeDelta, CallbackInfo ci) {
-		if (Zoom.isZooming() && ConfigHelper.getConfig(ConfigHelper.ConfigType.NORMAL, "zoom_mouse_mode").equals("scaled")) {
-			double mouseSensitivity = ClientData.CLIENT.options.getMouseSensitivity().getValue() * 0.6000000238418579 + 0.20000000298023224;
-			double mouseMultiplier = ((mouseSensitivity * mouseSensitivity * mouseSensitivity) * 8.0) * Math.max(ConfigHelper.getConfig(ConfigHelper.ConfigType.NORMAL, "zoom_transition").equals("smooth") ? MathHelper.lerp(timeDelta, Zoom.prevZoomMultiplier, Zoom.zoomMultiplier) : Zoom.getZoomMultiplier(), 0.001);
-			this.cursorDeltaX = this.cursorDeltaX * mouseMultiplier;
-			this.cursorDeltaY = this.cursorDeltaY * mouseMultiplier;
+		if (Zoom.isZooming() && ConfigHelper.getConfig(ConfigHelper.ConfigType.NORMAL, "zoom_scale_mode").equals("scaled")) {
+			this.cursorDeltaX *= Math.max(Zoom.zoomMultiplier, 0.001);
+			this.cursorDeltaY *= Math.max(Zoom.zoomMultiplier, 0.001);
 		}
 	}
 }